+++ punchbowl/level1/vignette.py
@@ -98,9 +98,39 @@ def correct_vignetting_task(data_object: NDCube, vignetting_path: str | pathlib.
     return data_object
 
 
-def generate_vignetting_calibration(path_vignetting: str, path_mask: str,
-                                    spacecraft: str) -> np.ndarray:
-    """Create calibration data for vignetting."""
+def generate_vignetting_calibration(path_vignetting: str,
+                                    path_mask: str,
+                                    spacecraft: str,
+                                    vignetting_threshold: float = 1.2,
+                                    rows_ignore: tuple = (13,15),
+                                    rows_adjust: tuple = (15,16),
+                                    rows_adjust_source: tuple = (16,20)) -> np.ndarray:
+    """
+    Create calibration data for vignetting.
+
+    Parameters
+    ----------
+    path_vignetting : str
+        path to raw input vignetting function
+    path_mask : str
+        path to spacecraft mask function
+    spacecraft : str
+        spacecraft number
+    vignetting_threshold : float, optional
+        threshold for bad vignetting pixels, by default 1.2
+    rows_ignore : tuple, optional
+        rows to exclude entirely from original vignetting data, by default (13,15) for 128x128 input
+    rows_adjust : tuple, optional
+        rows to adjust to the minimum of a set of rows above (per column), by default (15,16) for 128x128 input
+    rows_adjust_source : tuple, optional
+        rows to use for statistics to adjust vignetting rows as above, by default (16,20) for 128x128 input
+
+    Returns
+    -------
+    np.ndarray
+        vignetting function array
+
+    """
     if spacecraft in ["1", "2", "3"]:
         with open(path_vignetting) as f:
             lines = f.readlines()
@@ -115,27 +145,31 @@ def generate_vignetting_calibration(path_vignetting: str, path_mask: str,
         bin_size = int(bin_size)
 
         values = np.array([float(v) for line in lines[1:] for v in line.split()])
-        vignetting_original = values[:num_bins**2].reshape((num_bins, num_bins))
+        vignetting = values[:num_bins**2].reshape((num_bins, num_bins))
 
-        vignetting_original[np.isnan(vignetting_original)] = 0
+        vignetting[vignetting > vignetting_threshold] = np.nan
 
-        vignetting_original[6,:] = 0
+        vignetting[rows_ignore[0]:rows_ignore[1],:] = np.nan
+        vignetting[rows_adjust[0]:rows_adjust[1],:] = np.min(vignetting[rows_adjust_source[0]:rows_adjust_source[1],:],
+                                                             axis=0)
 
         wcs_vignetting = WCS(naxis=2)
 
         wcs_wfi = WCS(naxis=2)
-        wcs_wfi.wcs.cdelt = wcs_wfi.wcs.cdelt * vignetting_original.shape[0] / 2048.
+        wcs_wfi.wcs.cdelt = wcs_wfi.wcs.cdelt * vignetting.shape[0] / 2048.
 
-        vignetting = reproject_adaptive((vignetting_original, wcs_vignetting),
-                                        shape_out=(2048,2048),
-                                        output_projection=wcs_wfi,
-                                        boundary_mode="ignore")[0]
+        vignetting_reprojected = reproject_adaptive((vignetting, wcs_vignetting),
+                                                shape_out=(2048,2048),
+                                                output_projection=wcs_wfi,
+                                                boundary_mode="ignore",
+                                                bad_value_mode="ignore",
+                                                return_footprint=False)
 
-        vignetting = vignetting * mask
+        vignetting_reprojected = vignetting_reprojected * mask
 
-        vignetting[mask == 0] = 1
+        vignetting_reprojected[mask == 0] = 1
 
-        return vignetting
+        return vignetting_reprojected
     if spacecraft=="4":
         # TODO: implement NFI speckle inclusion
         return np.ones((2048,2048))